refactor(diagonalize): new submodule
Pulled out the get_principal_axes function into a new submodule "diagonalize".
Also moved over a couple inertia functions.
Similarly separated out the test functions that exist thus far.
